fix : dailyList
